+++ src/main/java/me/pajic/mapstitch/item/AtlasItem.java
@@ -278,7 +278,7 @@ public class AtlasItem extends Item {
 					int distX = Math.abs(mapData.centerX - posX);
 					int distZ = Math.abs(mapData.centerZ - posZ);
 					int scale = Math.powExact(2, mapData.scale);
-					if (distX < 64 * scale && distZ < 64 * scale) {
+					if (distX <= 64 * scale && distZ <= 64 * scale) {
 						atlas.set(ModDataComponents.ATLAS_ACTIVE_MAP_ID, mapId.id());
 						return;
 					}