+++ src/main/java/me/pajic/mapstitch/platform/neoforge/NeoforgeClientEventSubscriber.java
@@ -38,7 +38,16 @@ public class NeoforgeClientEventSubscriber {
 	@SubscribeEvent
 	private static void initKeybinds(RegisterKeyMappingsEvent event) {
 		event.registerCategory(ModKeybinds.MOD_KEYS);
+		event.registerCategory(ModKeybinds.MOD_KEYS_WORLD_MAP);
 		event.register(ModKeybinds.OPEN_WORLD_MAP);
+		event.register(ModKeybinds.SCALE_UP);
+		event.register(ModKeybinds.SCALE_DOWN);
+		event.register(ModKeybinds.DIMENSION_UP);
+		event.register(ModKeybinds.DIMENSION_DOWN);
+		event.register(ModKeybinds.FOLLOW_PLAYER);
+		event.register(ModKeybinds.TOGGLE_GRID);
+		event.register(ModKeybinds.EJECT_MAP);
+		event.register(ModKeybinds.TOGGLE_HELP);
 	}
 
 	@SubscribeEvent