+++ src/main/java/me/pajic/mapstitch/recipe/ModRecipes.java
@@ -3,7 +3,7 @@ package me.pajic.mapstitch.recipe;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 
 public class ModRecipes {
-	public static RecipeSerializer<AtlasRecipe> ATLAS = new RecipeSerializer<>(AtlasRecipe.MAP_CODEC, AtlasRecipe.STREAM_CODEC);
+	public static final RecipeSerializer<AtlasRecipe> ATLAS = new RecipeSerializer<>(AtlasRecipe.MAP_CODEC, AtlasRecipe.STREAM_CODEC);
 
 	public static void init() {}
 }