+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-- Pressing Ctrl + Q when hovering over a map in the world map will now drop it. If the map is inside an atlas, it will be ejected from it before being dropped.
-- Picking up an atlas no longer resets the eject mode it's set to.
+- Added a row of control buttons to the world map screen as an alternative to using keybinds. The buttons can be turned off in the config.
+- Most world map controls can now be rebound in the vanilla Controls menu.
+- Changed default keybinds to not use the arrow keys, in order to allow navigating the buttons using arrow keys:
+  - Next scale: Up arrow key → S
+  - Previous scale: Down arrow key → Unbound
+  - Next dimension: Right arrow key → D
+  - Previous scale: Left arrow key → Unbound